requirements.txt
numpy>=1.20.0
opencv_python>=4.7.0.68
pyspellchecker>=0.7.3
tensorflow>=2.11.0
tensorflow_intel>=2.11.0
